requirements.txt
pygame-ce>=2.0.0
Cython
setuptools

